Change options parser and add one more option and refactor
The OptionsParser is slightly better with the
argument being the directory or file.
It seems odd to use an option (--path) for a required input.
I also refactored the print into functions and created a main
function (best practice).
In addition, the program can be called as an executable as well now
and works both ways.